server/pyproject.toml
[tool.poetry]
name = "nmdbnb-server"
version = "0.1.0"
description = ""
authors = ["Paolo Chang <chanhwan.chang@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nmdbnb_server"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
